+++ backend-rust/src/graphql_api/search_result.rs
@@ -1,10 +1,14 @@
 use super::{
-    baker, block::Block, contract, get_config, get_pool, todo_api, token, ApiError, ApiResult,
-    ConnectionQuery,
+    baker::{self, Baker},
+    block::Block,
+    contract::{self, Contract, ContractSnapshot},
+    get_config, get_pool, todo_api, token, ApiError, ApiResult, ConnectionQuery,
 };
 use crate::{
     connection::DescendingI64,
-    graphql_api::{account::Account, node_status::NodeStatus, transaction::Transaction},
+    graphql_api::{
+        account::Account, baker::CurrentBaker, node_status::NodeStatus, transaction::Transaction,
+    },
     transaction_event::Event,
     transaction_reject::TransactionRejectReason,
     transaction_type::{
@@ -12,13 +16,13 @@ use crate::{
         UpdateTransactionType,
     },
 };
-use async_graphql::{connection, Context, Object};
+use async_graphql::{
+    connection::{self, CursorType},
+    Context, Object,
+};
 use futures::TryStreamExt;
 use regex::Regex;
-use std::{
-    cmp::{max, min},
-    str::FromStr,
-};
+use std::str::FromStr;
 
 pub struct SearchResult {
     pub query: String,
@@ -28,30 +32,146 @@ pub struct SearchResult {
 impl SearchResult {
     async fn contracts<'a>(
         &self,
-        _ctx: &Context<'a>,
-        #[graphql(desc = "Returns the first _n_ elements from the list.")] _first: Option<i32>,
+        ctx: &Context<'a>,
+        #[graphql(desc = "Returns the first _n_ elements from the list.")] first: Option<u64>,
         #[graphql(desc = "Returns the elements in the list that come after the specified cursor.")]
-        _after: Option<String>,
-        #[graphql(desc = "Returns the last _n_ elements from the list.")] _last: Option<i32>,
+        after: Option<String>,
+        #[graphql(desc = "Returns the last _n_ elements from the list.")] last: Option<u64>,
         #[graphql(desc = "Returns the elements in the list that come before the specified cursor.")]
-        _before: Option<String>,
+        before: Option<String>,
     ) -> ApiResult<connection::Connection<String, contract::Contract>> {
-        todo_api!()
+        let contract_index_regex: Regex = Regex::new("^[0-9]+$")
+            .map_err(|e| ApiError::InternalError(format!("Invalid regex: {}", e)))?;
+        let pool = get_pool(ctx)?;
+        let config = get_config(ctx)?;
+        let query = ConnectionQuery::<i64>::new(
+            first,
+            after,
+            last,
+            before,
+            config.contract_connection_limit,
+        )?;
+        let mut connection = connection::Connection::new(false, false);
+        if !contract_index_regex.is_match(&self.query) {
+            return Ok(connection);
+        }
+        let lower_case_query = self.query.to_lowercase();
+
+        let mut row_stream = sqlx::query!(
+            "SELECT * FROM (
+                SELECT
+                    contracts.index as index,
+                    sub_index,
+                    module_reference,
+                    name as contract_name,
+                    contracts.amount,
+                    blocks.slot_time as block_slot_time,
+                    transactions.block_height,
+                    transactions.hash as transaction_hash,
+                    accounts.address as creator
+                FROM contracts
+                    JOIN transactions ON transaction_index = transactions.index
+                    JOIN blocks ON transactions.block_height = blocks.height
+                    JOIN accounts ON transactions.sender_index = accounts.index
+                WHERE 
+                    contracts.index = $5 AND
+                    contracts.index > $1 AND 
+                    contracts.index < $2
+                ORDER BY
+                    (CASE WHEN $4 THEN contracts.index END) ASC,
+                    (CASE WHEN NOT $4 THEN contracts.index END) DESC
+                LIMIT $3
+            ) AS contract_data
+            ORDER BY contract_data.index DESC",
+            i64::from(query.from),
+            i64::from(query.to),
+            query.limit,
+            query.is_last,
+            lower_case_query.parse::<i64>().ok(),
+        )
+        .fetch(pool);
+
+        while let Some(row) = row_stream.try_next().await? {
+            let contract_address_index =
+                row.index.try_into().map_err(|e: <u64 as TryFrom<i64>>::Error| {
+                    ApiError::InternalError(e.to_string())
+                })?;
+            let contract_address_sub_index =
+                row.sub_index.try_into().map_err(|e: <u64 as TryFrom<i64>>::Error| {
+                    ApiError::InternalError(e.to_string())
+                })?;
+
+            let snapshot = ContractSnapshot {
+                block_height: row.block_height,
+                contract_address_index,
+                contract_address_sub_index,
+                contract_name: row.contract_name,
+                module_reference: row.module_reference,
+                amount: row.amount.try_into()?,
+            };
+
+            let contract = Contract {
+                contract_address_index,
+                contract_address_sub_index,
+                contract_address: format!(
+                    "<{},{}>",
+                    contract_address_index, contract_address_sub_index
+                ),
+                creator: row.creator.into(),
+                block_height: row.block_height,
+                transaction_hash: row.transaction_hash,
+                block_slot_time: row.block_slot_time,
+                snapshot,
+            };
+
+            connection
+                .edges
+                .push(connection::Edge::new(contract.contract_address_index.to_string(), contract));
+        }
+
+        if let (Some(page_min_id), Some(page_max_id)) =
+            (connection.edges.first(), connection.edges.last())
+        {
+            let result = sqlx::query!(
+                "
+                    SELECT MAX(index) as db_max_index, MIN(index) as db_min_index
+                    FROM contracts
+                    WHERE contracts.index = $1
+                ",
+                lower_case_query.parse::<i64>().ok()
+            )
+            .fetch_one(pool)
+            .await?;
+
+            let page_max: i64 =
+                page_max_id.node.contract_address_index.0.try_into().map_err(|e| {
+                    ApiError::InternalError(format!("A contract index is too large: {}", e))
+                })?;
+            let page_min: i64 =
+                page_min_id.node.contract_address_index.0.try_into().map_err(|e| {
+                    ApiError::InternalError(format!("A contract index is too large: {}", e))
+                })?;
+
+            connection.has_previous_page =
+                result.db_max_index.map_or(false, |db_max| db_max > page_max);
+            connection.has_next_page =
+                result.db_min_index.map_or(false, |db_min| db_min < page_min);
+        }
+
+        Ok(connection)
     }
 
-    //    async fn modules(
-    //        &self,
-    //        #[graphql(desc = "Returns the first _n_ elements from the list.")]
-    // _first: Option<i32>,        #[graphql(desc = "Returns the elements in the
-    //     list that come after the specified cursor.")]
-    //        _after: Option<String>,
-    //        #[graphql(desc = "Returns the last _n_ elements from the list.")]
-    // _last: Option<i32>,        #[graphql(desc = "Returns the elements in the
-    // list that come before the     specified cursor.")]
-    //        _before: Option<String>,
-    //    ) -> ApiResult<connection::Connection<String, Module>> {
-    //        todo_api!()
-    //    }
+    // async fn modules(
+    //     &self,
+    //     #[graphql(desc = "Returns the first _n_ elements from the list.")]
+    // _first: Option<i32>,     #[graphql(desc = "Returns the elements in the
+    // list that come after the specified cursor.")]     _after: Option<String>,
+    //     #[graphql(desc = "Returns the last _n_ elements from the list.")] _last:
+    // Option<i32>,     #[graphql(desc = "Returns the elements in the list that
+    // come before the     specified cursor.")]     _before: Option<String>,
+    // ) -> ApiResult<connection::Connection<String, Module>> {
+    //     todo_api!()
+    // }
 
     async fn blocks(
         &self,
@@ -105,22 +225,13 @@ impl SearchResult {
         )
         .fetch(pool);
 
-        let mut min_height = None;
-        let mut max_height = None;
         while let Some(block) = rows.try_next().await? {
-            min_height = Some(match min_height {
-                None => block.height,
-                Some(current_min) => min(current_min, block.height),
-            });
-
-            max_height = Some(match max_height {
-                None => block.height,
-                Some(current_max) => max(current_max, block.height),
-            });
             connection.edges.push(connection::Edge::new(block.height.to_string(), block));
         }
 
-        if let (Some(page_min_height), Some(page_max_height)) = (min_height, max_height) {
+        if let (Some(page_min_height), Some(page_max_height)) =
+            (connection.edges.first(), connection.edges.last())
+        {
             let result = sqlx::query!(
                 r#"
                     SELECT MAX(height) as max_height, MIN(height) as min_height
@@ -134,10 +245,11 @@ impl SearchResult {
             )
             .fetch_one(pool)
             .await?;
+
             connection.has_previous_page =
-                result.max_height.map_or(false, |db_max| db_max > page_max_height);
+                result.max_height.map_or(false, |db_max| db_max > page_max_height.node.height);
             connection.has_next_page =
-                result.min_height.map_or(false, |db_min| db_min < page_min_height);
+                result.min_height.map_or(false, |db_min| db_min < page_min_height.node.height);
         }
         Ok(connection)
     }
@@ -309,22 +421,13 @@ impl SearchResult {
         .fetch_all(pool)
         .await?;
 
-        let mut min_index = None;
-        let mut max_index = None;
         for account in accounts {
-            min_index = Some(match min_index {
-                None => account.index,
-                Some(current_min) => min(current_min, account.index),
-            });
-
-            max_index = Some(match max_index {
-                None => account.index,
-                Some(current_max) => max(current_max, account.index),
-            });
             connection.edges.push(connection::Edge::new(account.index.to_string(), account));
         }
 
-        if let (Some(page_min_id), Some(page_max_id)) = (min_index, max_index) {
+        if let (Some(page_min_id), Some(page_max_id)) =
+            (connection.edges.first(), connection.edges.last())
+        {
             let result = sqlx::query!(
                 r#"
                     SELECT MAX(index) as max_id, MIN(index) as min_id
@@ -338,22 +441,114 @@ impl SearchResult {
             .await?;
 
             connection.has_previous_page =
-                result.min_id.map_or(false, |db_min| db_min < page_min_id);
-            connection.has_next_page = result.max_id.map_or(false, |db_max| db_max > page_max_id);
+                result.min_id.map_or(false, |db_min| db_min < page_min_id.node.index);
+            connection.has_next_page =
+                result.max_id.map_or(false, |db_max| db_max > page_max_id.node.index);
         }
         Ok(connection)
     }
 
     async fn bakers(
         &self,
-        #[graphql(desc = "Returns the first _n_ elements from the list.")] _first: Option<i32>,
+        ctx: &Context<'_>,
+        #[graphql(desc = "Returns the first _n_ elements from the list.")] first: Option<u64>,
         #[graphql(desc = "Returns the elements in the list that come after the specified cursor.")]
-        _after: Option<String>,
-        #[graphql(desc = "Returns the last _n_ elements from the list.")] _last: Option<i32>,
+        after: Option<String>,
+        #[graphql(desc = "Returns the last _n_ elements from the list.")] last: Option<u64>,
         #[graphql(desc = "Returns the elements in the list that come before the specified cursor.")]
-        _before: Option<String>,
+        before: Option<String>,
     ) -> ApiResult<connection::Connection<String, baker::Baker>> {
-        todo_api!()
+        let baker_index_regex: Regex = Regex::new("^[0-9]+$")
+            .map_err(|e| ApiError::InternalError(format!("Invalid regex: {}", e)))?;
+        let pool = get_pool(ctx)?;
+        let config = get_config(ctx)?;
+        let query =
+            ConnectionQuery::<i64>::new(first, after, last, before, config.baker_connection_limit)?;
+        let mut connection = connection::Connection::new(false, false);
+        if !baker_index_regex.is_match(&self.query) {
+            return Ok(connection);
+        }
+        let lower_case_query = self.query.to_lowercase();
+
+        let mut row_stream = sqlx::query_as!(
+            CurrentBaker,
+            r#"SELECT * FROM (
+                SELECT
+                    bakers.id AS id,
+                    staked,
+                    restake_earnings,
+                    open_status as "open_status: _",
+                    metadata_url,
+                    self_suspended,
+                    inactive_suspended,
+                    primed_for_suspension,
+                    transaction_commission,
+                    baking_commission,
+                    finalization_commission,
+                    payday_transaction_commission as "payday_transaction_commission?",
+                    payday_baking_commission as "payday_baking_commission?",
+                    payday_finalization_commission as "payday_finalization_commission?",
+                    payday_lottery_power as "payday_lottery_power?",
+                    payday_ranking_by_lottery_powers as "payday_ranking_by_lottery_powers?",
+                    (SELECT MAX(payday_ranking_by_lottery_powers) FROM bakers_payday_lottery_powers) as "payday_total_ranking_by_lottery_powers?",
+                    pool_total_staked,
+                    pool_delegator_count,
+                    baker_apy,
+                    delegators_apy
+                FROM bakers
+                    LEFT JOIN latest_baker_apy_30_days
+                        ON latest_baker_apy_30_days.id = bakers.id
+                    LEFT JOIN bakers_payday_commission_rates
+                        ON bakers_payday_commission_rates.id = bakers.id
+                    LEFT JOIN bakers_payday_lottery_powers
+                        ON bakers_payday_lottery_powers.id = bakers.id
+                WHERE
+                    bakers.id = $5 AND
+                    (bakers.id > $1 AND 
+                    bakers.id < $2)
+                ORDER BY
+                    (CASE WHEN $3     THEN bakers.id END) DESC,
+                    (CASE WHEN NOT $3 THEN bakers.id END) ASC
+                LIMIT $4
+            ) ORDER BY id ASC"#,
+            query.from,                                        // $1
+            query.to,                                          // $2
+            query.is_last,                                     // $3
+            query.limit,                                       // $4
+            lower_case_query.parse::<i64>().ok()               // $5
+        )
+        .fetch(pool);
+        while let Some(row) = row_stream.try_next().await? {
+            let cursor = row.id.encode_cursor();
+            connection.edges.push(connection::Edge::new(cursor, Baker::Current(Box::new(row))));
+        }
+
+        let (Some(first_item), Some(last_item)) =
+            (connection.edges.first(), connection.edges.last())
+        else {
+            // No items so we just return.
+            return Ok(connection);
+        };
+
+        {
+            let bounds = sqlx::query!(
+                "SELECT
+                    MAX(id),
+                    MIN(id)
+                FROM bakers
+                WHERE
+                    bakers.id = $1
+                ",
+                lower_case_query.parse::<i64>().ok()
+            )
+            .fetch_one(pool)
+            .await?;
+            if let (Some(min), Some(max)) = (bounds.min, bounds.max) {
+                connection.has_previous_page = min < first_item.node.get_id();
+                connection.has_next_page = max > last_item.node.get_id();
+            }
+        }
+        Ok(connection)
     }
 
     async fn node_statuses(